pyproject.toml
[tool.poetry]
name = "pokemon-vgc-assistant"
version = "0.1.0"
description = ""
authors = ["Felipe Kummer <felipe.s.kummer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
apache-airflow = "^2.8.0"
psycopg2-binary = "^2.9.3"
pendulum = "^2"
docutils = ">=0.16,<0.21"
google-cloud-storage = "^2.18.2"
apache-airflow-providers-google = "^10.22.0"

[tool.poetry.dev-dependencies]
ipython = "^8.4.0"
black = "^22.3.0"
isort = "^5.10.1"
PyYAML = "^6.0"
dockerfile = "^3.2.0"
pandas = ">=2.1.2,<2.2"
requests = "^2.28.1"
responses = "^0.22.0"
#pytest-mock = "3.2.0"
#pytest-responses = "0.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 160
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 160
multi_line_output = 3
include_trailing_comma = true

[tool.pytest.ini_options]
markers = ["optional: not required, nor saved in test_output.txt"]
